Drive System
I have updated the DriveSystem class and changed the main robot
template to be compatible with the new Drive System. I also created a
gyro in the RobotTemplate and added in a variable called drive mode to
allow the user to choose between driving with the gyro or driving
without it.