Toggle Logic READ THIS EVERYONE
I started on the Toggle logic.
Alex, the way the Catch stuff is setup in the ball handler, I can't
toggle catch mode on or off. You have a "catch" method that is similar
to the shoot method. What we really need is a "enableCatchMode" and
"disableCatchMode" method.
I probably won't be there tonight, tomorrow night, or Saturday...
But if you keep me informed (Zephan) I can keep working on the
RobotTemplate and RangeFinder stuff. :-)
Thanks. :-)